% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_inv.R
\name{All_inv}
\alias{All_inv}
\title{Inverse rearrangement algorithm on all modes}
\usage{
All_inv(ten_bundle, mode = "all")
}
\arguments{
\item{ten_bundle}{A list of two objects, a multi-dimensional array and a list of indices along the modes to reverse}

\item{mode}{A vector implying the modes to reverse, correspondent to the list in 'ten_bundle', set as 'all' by default}
}
\value{
A multi-dimensional array
}
\description{
The inverse algorithm of rearrangement algorithm for all modes
}
