% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_rearrange.R
\name{All_rearrange}
\alias{All_rearrange}
\title{Rearrangement algorithm on all modes}
\usage{
All_rearrange(ten, except = "NA", key = "NA")
}
\arguments{
\item{ten}{A multi-dimensional array}

\item{except}{A vector of integers implying the modes not to rearrange, set as 'NA' by default}

\item{key}{The value to which rearrange is according, set as 'NA' by default}
}
\value{
A list of two objects, a rearranged tensor and a list of indices rearranged (ordered by the rearranged modes)
}
\description{
Rearrangement algorithm for all modes
}
\details{
A rearrangement algorithm on higher order tensor to rearrange all missing entries to a corner block. The case of mode-2 tensor returns to the work of Bai and Ng in 2021.
}
