% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_trans.R
\name{H_trans}
\alias{H_trans}
\title{Mode-k Block-a Block-b transformation matrix}
\usage{
H_trans(A, B)
}
\arguments{
\item{A}{A matrix with n rows and l columns}

\item{B}{A matrix with k rows and l columns, with k no larger than n}
}
\value{
A matrix
}
\description{
Defined for an ongoing working paper
}
