% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_ind.R
\name{obs_ind}
\alias{obs_ind}
\title{Largest index observed}
\usage{
obs_ind(ten, k)
}
\arguments{
\item{ten}{A multi-dimensional array}

\item{k}{An integer specifying the tensor mode to check}
}
\value{
An integer specifying the largest possible index
}
\description{
Finding the largest index I along a tensor mode k such that the block until index I contains no NA
}
