% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange.R
\name{rearrange}
\alias{rearrange}
\title{Rearrangement algorithm on mode k}
\usage{
rearrange(ten, k, key = "NA")
}
\arguments{
\item{ten}{A multi-dimensional array}

\item{k}{An integer specifying the mode to arrange}

\item{key}{The value to which rearrange is according, set as 'NA' by default}
}
\value{
A list of two objects, a rearranged tensor and the indices rearranged
}
\description{
Rearrangement algorithm on one mode
}
\details{
A rearrangement algorithm on higher order tensor to rearrange all missing entries along mode k to the end.
}
