% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange_inv.R
\name{rearrange_inv}
\alias{rearrange_inv}
\title{Inverse rearrangement algorithm on mode k}
\usage{
rearrange_inv(ten, k, l)
}
\arguments{
\item{ten}{A multi-dimensional array}

\item{k}{An integer specifying the mode k to rearrange backwards}

\item{l}{A vector specifying the original indices rearranged}
}
\value{
A multi-dimensional array
}
\description{
The inverse algorithm of rearrangement algorithm for one mode
}
